requirements.txt
asgiref==3.2.10
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.1
black==22.1.0
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-cors-headers==3.11.0
djangorestframework==3.12.4
entrypoints==0.4
env==0.1.0
executing==0.8.2
filelock==3.3.2
gunicorn==20.1.0
ipykernel==6.9.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
nest-asyncio==1.5.4
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.0
prompt-toolkit==3.0.27
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
static3==0.7.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
virtualenv==20.10.0
waitress==2.1.1
wcwidth==0.2.5
whitenoise==6.0.0
